sandbox: Move the services into the `apiserver.services` package.
