+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/LedgerApiServer.scala
@@ -12,7 +12,6 @@ import akka.stream.ActorMaterializer
 import com.codahale.metrics.MetricRegistry
 import com.digitalasset.grpc.adapter.{AkkaExecutionSequencerPool, ExecutionSequencerFactory}
 import com.digitalasset.platform.common.logging.NamedLoggerFactory
-import com.digitalasset.platform.sandbox.metrics.MetricsInterceptor
 import io.grpc.ServerInterceptor
 import io.grpc.netty.NettyServerBuilder
 import io.netty.channel.nio.NioEventLoopGroup
