+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/services/ApiLedgerConfigurationService.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2019 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.sandbox.services
+package com.digitalasset.platform.apiserver.services
 
 import akka.NotUsed
 import akka.stream.Materializer
@@ -15,9 +15,9 @@ import com.digitalasset.platform.api.grpc.GrpcApiService
 import com.digitalasset.platform.common.logging.NamedLoggerFactory
 import com.digitalasset.platform.common.util.DirectExecutionContext
 import com.digitalasset.platform.server.api.validation.LedgerConfigurationServiceValidation
-
 import io.grpc.{BindableService, ServerServiceDefinition}
 import org.slf4j.Logger
+
 import scala.concurrent.ExecutionContext
 
 class ApiLedgerConfigurationService private (configurationService: IndexConfigurationService)(
