+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/services/admin/ApiPartyManagementService.scala
@@ -1,13 +1,13 @@
 // Copyright (c) 2019 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.sandbox.services.admin
+package com.digitalasset.platform.apiserver.services.admin
 
 import java.util.UUID
 
 import akka.actor.Scheduler
-import akka.stream.{ActorMaterializer, Materializer}
 import akka.stream.scaladsl.Sink
+import akka.stream.{ActorMaterializer, Materializer}
 import com.daml.ledger.participant.state.index.v2.{
   IndexPartyManagementService,
   IndexTransactionsService
@@ -17,8 +17,8 @@ import com.daml.ledger.participant.state.v1.{SubmissionId, SubmissionResult, Wri
 import com.digitalasset.daml.lf.data.Ref
 import com.digitalasset.grpc.adapter.ExecutionSequencerFactory
 import com.digitalasset.ledger.api.domain
-import com.digitalasset.ledger.api.domain.{LedgerOffset, PartyEntry}
 import com.digitalasset.ledger.api.domain.PartyEntry.{AllocationAccepted, AllocationRejected}
+import com.digitalasset.ledger.api.domain.{LedgerOffset, PartyEntry}
 import com.digitalasset.ledger.api.v1.admin.party_management_service.PartyManagementServiceGrpc.PartyManagementService
 import com.digitalasset.ledger.api.v1.admin.party_management_service._
 import com.digitalasset.platform.api.grpc.GrpcApiService
