+++ ledger/sandbox/src/test/suite/scala/com/digitalasset/platform/apiserver/services/admin/PollingUtilsSpec.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2019 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.sandbox.services.admin
+package com.digitalasset.platform.apiserver.services.admin
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
